Make curl command and variables configurable.
Introduce g:dict_curl_command and g:dict_curl_options to allow for customization. For example, without direct access to the Internet, I need to add "--socks5 socks-server" (because curl doesn't pick up the http_proxy environment variable when using the dict protocol).
Some users may not have the curl command on their PATH, so having that configurable is helpful, too. Note that I didn't use shellescape() on it, because that function isn't available in Vim 7.0/1, the default doesn't require it, and users customizing it can add that themselves.